Move health to service layer
